requirements.txt
beautifulsoup4>=4.11.1
Django>=4.1
requests>=2.28.1
igdb-api-v4==0.0.5
djangorestframework==3.14.0
